refactor(ming-tts): clean code
