+++ sglang_omni/models/ming_tts/ar_runtime.py
@@ -1,10 +1,14 @@
 # SPDX-License-Identifier: Apache-2.0
-"""Ming-Omni-TTS AR state contracts.
+"""Ming-Omni-TTS AR runtime contracts.
 
-The Ming-TTS AR loop is not a normal token-logits-sampler loop.  The
-SGLang-managed backbone produces hidden states, and a Ming-specific recurrent
-state machine turns those hidden states into continuous acoustic latents and
-feedback embeddings.
+This module owns request-local AR recurrence state and fixed-address feedback
+buffers shared by request builders, the model runner, and result serialization.
+Cross-stage serializable payload fields belong in payload_types.py; AR backbone
+math and weights belong in sglang_model.py/model_runner.py.
+
+The Ming-TTS AR loop is not a normal token-logits-sampler loop. The
+SGLang-managed backbone produces hidden states, and the Ming AR tail turns
+those hidden states into continuous acoustic latents and feedback embeddings.
 
 Current step contract:
 
@@ -17,6 +21,7 @@ decode before forward:
   input: pending feedback embedding per active request
   output: fixed feedback buffer updated in-place
   side state: input_ids rewritten to row ids
+  graph padding rows: intentionally untouched and ignored by SGLang metadata
 
 backbone forward:
   input: positions, KV cache, row feedback embedding
@@ -41,10 +46,9 @@ import torch
 class MingARRequestState:
     """Per-request Ming AR recurrence state.
 
-    This mirrors the legacy fields on ``MingTTSSGLangRequestData`` during the
-    first refactor stage.  Callers should use ``get_ming_ar_state`` and
-    ``sync_ming_ar_state_to_legacy`` so old and new storage stay compatible
-    while the runner is migrated.
+    This object is the single owner of Ming AR recurrence state.  Scheduler
+    request data may keep generic counters for SGLang bookkeeping, but Ming
+    generation must not read AR state from those generic fields.
     """
 
     generation_steps: int = 0
@@ -52,7 +56,6 @@ class MingARRequestState:
     cfg: float = 2.0
     sigma: float = 0.25
     flow_temperature: float = 0.0
-    seed: int | None = None
     audio_patch_token_id: int = 0
     audio_eos_token_id: int = 0
     audio_token_id: int = 0
@@ -63,48 +66,6 @@ class MingARRequestState:
     generated_last_chunk: list[bool] = field(default_factory=list)
     stop_step: int | None = None
     engine_start_s: float = 0.0
-
-    @classmethod
-    def from_legacy(cls, data: Any) -> "MingARRequestState":
-        queue = getattr(data, "pending_feedback_queue", None)
-        if queue is None:
-            queue = collections.deque()
-
-        generated_latents = getattr(data, "generated_latents", None)
-        if generated_latents is None:
-            generated_latents = []
-
-        generated_last_chunk = getattr(data, "generated_last_chunk", None)
-        if generated_last_chunk is None:
-            generated_last_chunk = []
-
-        return cls(
-            generation_steps=int(getattr(data, "generation_steps", 0) or 0),
-            max_decode_steps=int(getattr(data, "max_decode_steps", 0) or 0),
-            cfg=float(getattr(data, "cfg", 2.0)),
-            sigma=float(getattr(data, "sigma", 0.25)),
-            flow_temperature=float(getattr(data, "flow_temperature", 0.0)),
-            seed=getattr(data, "seed", None),
-            audio_patch_token_id=int(getattr(data, "audio_patch_token_id", 0) or 0),
-            audio_eos_token_id=int(getattr(data, "audio_eos_token_id", 0) or 0),
-            audio_token_id=int(getattr(data, "audio_token_id", 0) or 0),
-            prompt_latent_for_history=getattr(
-                data,
-                "prompt_latent_for_history",
-                None,
-            ),
-            latent_history=getattr(data, "latent_history", None),
-            pending_feedback_queue=queue,
-            generated_latents=generated_latents,
-            generated_last_chunk=generated_last_chunk,
-            stop_step=getattr(data, "stop_step", None),
-            engine_start_s=float(getattr(data, "engine_start_s", 0.0) or 0.0),
-        )
-
-    def sync_generation_step_from_legacy(self, data: Any) -> None:
-        self.generation_steps = int(
-            getattr(data, "generation_steps", self.generation_steps) or 0
-        )
 
     def ensure_latent_history(
         self,
@@ -179,8 +140,13 @@ class MingARRequestState:
         self.generated_latents.clear()
 
 
-class MingARDeviceStatePool:
-    """Fixed-address device state visible to Ming AR CUDA graph replay."""
+class MingARFeedbackStatePool:
+    """Fixed-address active-row feedback buffer for Ming AR CUDA graph replay.
+
+    CUDA graph replay can use a larger captured bucket than the active request
+    count. This pool only writes active rows [0, batch_size); padded rows are
+    intentionally left untouched and must be ignored by SGLang forward metadata.
+    """
 
     def __init__(self, decode_input_embedding: Any) -> None:
         self.decode_input_embedding = decode_input_embedding
@@ -253,34 +219,8 @@ class MingARDeviceStatePool:
 def get_ming_ar_state(data: Any) -> MingARRequestState:
     ar_state = getattr(data, "ar_state", None)
     if not isinstance(ar_state, MingARRequestState):
-        ar_state = MingARRequestState.from_legacy(data)
-        setattr(data, "ar_state", ar_state)
-    ar_state.sync_generation_step_from_legacy(data)
-    sync_ming_ar_state_to_legacy(data, ar_state)
+        raise RuntimeError("Ming TTS request data is missing MingARRequestState")
     return ar_state
-
-
-def sync_ming_ar_state_to_legacy(
-    data: Any,
-    ar_state: MingARRequestState,
-) -> None:
-    data.generation_steps = int(ar_state.generation_steps)
-    data.max_decode_steps = int(ar_state.max_decode_steps)
-    data.cfg = float(ar_state.cfg)
-    data.sigma = float(ar_state.sigma)
-    data.flow_temperature = float(ar_state.flow_temperature)
-    data.seed = ar_state.seed
-    data.audio_patch_token_id = int(ar_state.audio_patch_token_id)
-    data.audio_eos_token_id = int(ar_state.audio_eos_token_id)
-    data.audio_token_id = int(ar_state.audio_token_id)
-    data.prompt_latent_for_history = ar_state.prompt_latent_for_history
-    data.latent_history = ar_state.latent_history
-    data.pending_feedback_queue = ar_state.pending_feedback_queue
-    data.generated_latents = ar_state.generated_latents
-    data.generated_last_chunk = ar_state.generated_last_chunk
-    data.stop_step = ar_state.stop_step
-    data.engine_start_s = float(ar_state.engine_start_s)
-    data.ar_state = ar_state
 
 
 def update_ming_ar_latent_history_(
@@ -307,34 +247,25 @@ def release_ming_ar_runtime_tensors(data: Any) -> None:
     ar_state = getattr(data, "ar_state", None)
     if isinstance(ar_state, MingARRequestState):
         ar_state.release_runtime_tensors()
-        sync_ming_ar_state_to_legacy(data, ar_state)
-        return
-
-    data.latent_history = None
-    data.prompt_latent_for_history = None
-    if hasattr(getattr(data, "pending_feedback_queue", None), "clear"):
-        data.pending_feedback_queue.clear()
-    data.generated_latents = []
 
 
-def get_ming_ar_device_state_pool(model: Any) -> MingARDeviceStatePool:
-    pool = getattr(model, "_ming_ar_device_state_pool", None)
-    if isinstance(pool, MingARDeviceStatePool):
+def get_ming_ar_feedback_state_pool(model: Any) -> MingARFeedbackStatePool:
+    pool = getattr(model, "_ming_ar_feedback_state_pool", None)
+    if isinstance(pool, MingARFeedbackStatePool):
         return pool
     embedding = getattr(model, "_decode_input_embedding", None)
     if embedding is None:
         raise RuntimeError("Ming TTS model is missing _decode_input_embedding")
-    pool = MingARDeviceStatePool(embedding)
-    setattr(model, "_ming_ar_device_state_pool", pool)
+    pool = MingARFeedbackStatePool(embedding)
+    setattr(model, "_ming_ar_feedback_state_pool", pool)
     return pool
 
 
 __all__ = [
-    "MingARDeviceStatePool",
+    "MingARFeedbackStatePool",
     "MingARRequestState",
-    "get_ming_ar_device_state_pool",
+    "get_ming_ar_feedback_state_pool",
     "get_ming_ar_state",
     "release_ming_ar_runtime_tensors",
-    "sync_ming_ar_state_to_legacy",
     "update_ming_ar_latent_history_",
 ]
